+++ Source/EventFlow.Autofac/Extensions/EventFlowOptionsAutofacExtensions.cs
@@ -22,18 +22,34 @@
 
 using System;
 using Autofac;
+using EventFlow.Aggregates;
 using EventFlow.Configuration.Registrations;
+using EventFlow.Configuration.Registrations.Services;
 
 namespace EventFlow.Autofac.Extensions
 {
     public static class EventFlowOptionsAutofacExtensions
     {
         public static EventFlowOptions UseAutofacContainerBuilder(
+            this EventFlowOptions eventFlowOptions)
+        {
+            return eventFlowOptions
+                .UseAutofacContainerBuilder(new ContainerBuilder());
+        }
+
+        public static EventFlowOptions UseAutofacContainerBuilder(
             this EventFlowOptions eventFlowOptions,
             ContainerBuilder containerBuilder)
         {
             return eventFlowOptions
                 .UseServiceRegistration(new AutofacServiceRegistration(containerBuilder));
+        }
+
+        public static EventFlowOptions UseAutofacAggregateRootFactory(
+            this EventFlowOptions eventFlowOptions)
+        {
+            return eventFlowOptions
+                .RegisterServices(f => f.Register<IAggregateFactory, AutofacAggregateRootFactory>(Lifetime.Singleton));
         }
 
         public static IContainer CreateContainer(