+++ Source/EventFlow/EventStores/IEventJsonSerializer.cs
@@ -31,6 +31,8 @@ namespace EventFlow.EventStores
             IAggregateEvent aggregateEvent,
             IEnumerable<KeyValuePair<string, string>> metadatas);
 
+        IDomainEvent Deserialize(string json, IMetadata metadata);
+
         IDomainEvent Deserialize(
             ICommittedDomainEvent committedDomainEvent);